feat: implement wakeup source tracking for power management (Item 110)
Add a wakeup event source tracking subsystem that allows device
drivers (RTC, USB, power button, NIC) to register as wakeup
sources, report wakeup events, and have their event counts
tracked. The PM suspend path can query active sources to defer
suspend when events are in-flight.

- src/include/wakeup.h: Public API with registration, event
  counting, active-state query, and source listing
- src/power/wakeup.c: Fixed-size array implementation with
  spinlock protection, init logging, and formatted source dump
- Makefile: add src/power/wakeup.c to build

Also fix -Werror format string warnings found by 'make check':
- cmd_bg.c: cast pid to unsigned int instead of unsigned long
- cmd_blkid.c: cast mb/sectors to unsigned int
- cmd_bunzip2.c: cast in_size/opos to unsigned int
- cmd_calc.c: use %lld/%llx for int64_t result
- cmd_capprof.c: cast pid to unsigned int, rc to int
- cmd_ccbuilder.c: remove unnecessary unsigned long casts
- cmd_chown.c: cast uid_val/gid_val to unsigned int

+++ src/shell/cmds/cmd_bg.c
@@ -58,12 +58,12 @@ void cmd_bg(const char *args) {
         return;
     }
     if (pgid && libc_killpg(pgid, 18) == 0) {
-        kprintf("[%u] Continued\n", (unsigned long)pid);
+        kprintf("[%u] Continued\n", (unsigned int)pid);
         return;
     }
     if (libc_kill(pid, 18) == 0) {
-        kprintf("[%u] Continued\n", (unsigned long)pid);
+        kprintf("[%u] Continued\n", (unsigned int)pid);
         return;
     }
-    kprintf("No such job: %u\n", (unsigned long)pid);
+    kprintf("No such job: %u\n", (unsigned int)pid);
 }
