+++ src/shell/cmds/cmd_bunzip2.c
@@ -96,5 +96,5 @@ void cmd_bunzip2(const char *args) {
     }
 
     libc_vfs_write(outpath, outbuf, opos);
-    kprintf("Decompressed %u -> %u bytes (%s)\n", (unsigned long)in_size, (unsigned long)opos, outpath);
+    kprintf("Decompressed %u -> %u bytes (%s)\n", (unsigned int)in_size, (unsigned int)opos, outpath);
 }
